[TASK-21] Fix test cases based on question builder restructures.
