+++ src/tests/components/DetermineQuestionType/emojis.js
